fix: improve device view performence
