+++ src/views/dashboard/settings/Settings.tsx
@@ -20,7 +20,6 @@ export const SettingsView = () => {
             Dark mode
           </Switch.Label>
         </span>
-
         <Switch
           checked={theme === THEME.DARK}
           onChange={(checked) =>
